requirements.txt
playwright==1.41.2
curl-cffi==0.5.10
pyyaml==6.0.1
discord-webhook==1.3.0
pandas==2.2.0
scipy==1.12.0
